backend/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
astroid==2.15.5
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
beir==1.0.1
blis==0.7.9
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
confection==0.0.4
contourpy==1.0.7
cryptography==41.0.3
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.7
datasets==2.12.0
defusedxml==0.7.1
dill==0.3.6
dnspython==2.3.0
easy-elasticsearch==0.0.9
ecdsa==0.18.0
elasticsearch==7.9.1
exceptiongroup==1.1.1
faiss-cpu==1.7.4
fastapi==0.95.1
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
fuzzywuzzy==0.18.0
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gpl==0.1.4
gql==3.4.1
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
httpie==3.2.1
httplib2==0.22.0
huggingface-hub==0.14.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kaleido==0.2.1
kiwisolver==1.4.4
langchain==0.0.221
langchainplus-sdk==0.0.19
langcodes==3.3.0
lazy-object-proxy==1.9.0
Levenshtein==0.21.0
lit==16.0.5
llama-index==0.6.38.post1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mccabe==0.7.0
mdurl==0.1.2
motor==3.1.2
mplfinance==0.12.9b7
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.5
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.0
passlib==1.7.4
pathy==0.10.1
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.15.0
pluggy==1.0.0
preshed==3.0.8
prettytable==3.8.0
protobuf==4.23.0
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pyecharts==2.0.3
Pygments==2.15.1
pylint==2.17.4
pymongo==4.3.3
PyMySQL==1.1.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.3.1
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python-Levenshtein==0.21.0
python-multipart==0.0.6
pytrec-eval==0.5
pytz==2023.3
PyYAML==6.0
rapidfuzz==3.0.0
regex==2023.5.5
requests==2.28.2
requests-toolbelt==0.10.1
responses==0.18.0
rich==13.3.4
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
simplejson==3.19.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.17
srsly==2.4.6
starlette==0.26.1
sympy==1.12
tenacity==8.2.2
thinc==8.1.10
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.11.8
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.2
triton==2.0.0
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
wasabi==1.1.1
wcwidth==0.2.6
wrapt==1.15.0
xxhash==3.2.0
yarl==1.9.2
